requirements.txt
discord.py
mee6-py-api